pyproject.toml
[project]
name = "earthquake-hazards"
version = "0.0.1"
description = "Data engineering pipeline to publish a dashboard about earthquakes across the world"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "geopandas>=0.13.2",
    "pandas>=2.0.3",
    "shapely>=2.0.6",
]

[dependency-groups]
dev = [
    "apache-airflow[google,postgres]>=2.10.4",
    "astronomer-cosmos>=1.8.2",
    "dbt-bigquery>=1.8.3",
    "dbt-core>=1.8.7",
    "pytest>=8.3.4",
]

[tool.pyright]
include = ["src", "tests"]
reportUnusedCallResult = false
reportExplicitAny = false

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
